stealthpilot-ai-interview-assistant-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
aiosqlite==0.19.0
anthropic==0.24.0
websockets==12.0
deepgram-sdk==3.5.0
openai==1.58.1
google-generativeai==0.3.2
httpx==0.27.0
